daemon: move code related to stats together
